Downgrade DRF to fix swagger UI issue
Newer versions of DRF cause swagger UI to use 'http' requests when using
'try it out' which breaks on a production deploy using 'https'. There is a
selection for 'schemes' but it only has 'http' as a choice. It's not clear
what is causing this issue and it is easier for now to just go back to the
working version.